Fix Xello file formatting issue
This line was trying to replace CRLF with LF line endings, but because I originally used single quotes instead of double quotes, the `n was not translated into LF, it was literally put into the document, wrecking the formatting for half of the file. With double quotes, this should translate into a LF character, and this issue should be fixed.